+++ src/Gnugat/SoulMeMaybe/NetSoulProtocol/Request/StartAuthenticationRequest.php
@@ -9,24 +9,16 @@ namespace Gnugat\SoulMeMaybe\NetSoulProtocol\Request;
  */
 class StartAuthenticationRequest extends AbstractRequest
 {
-    /**
-     * @var string The command name.
-     */
+    /** @var string The command name. */
     public $commandName = 'auth_ag';
 
-    /**
-     * @var string The authentication type.
-     */
+    /** @var string The authentication type. */
     public $authenticationType = 'ext_user';
 
-    /**
-     * @var string The unused argument 1.
-     */
+    /** @var string The unused argument 1. */
     public $firstUnusedArgument = 'none';
 
-    /**
-     * @var string The unused argument 2.
-     */
+    /** @var string The unused argument 2. */
     public $secondUnusedArgument = 'none';
 
     /**