+++ src/Gnugat/SoulMeMaybe/NetSoulProtocol/Request/StateRequest.php
@@ -9,23 +9,32 @@ namespace Gnugat\SoulMeMaybe\NetSoulProtocol\Request;
  */
 class StateRequest extends AbstractRequest
 {
-    /**
-     * @var string The command name.
-     */
+    /** @var array The states. */
+    public static $states = array(
+        'actif',
+        'away',
+        'connection',
+        'idle',
+        'lock',
+        'server',
+        'none',
+    );
+
+    /** @var string The command name. */
     public $commandName = 'state';
 
-    /**
-     * @var string The state and timestamp.
-     */
+    /** @var string The state and timestamp. */
     public $stateAndTimestamp;
 
     /**
      * The constructor.
+     *
+     * @param string $state The state.
      */
-    public function __construct()
+    public function __construct($state)
     {
         $this->stateAndTimestamp = implode(':', array(
-            'active',
+            $state,
             strval(time()),
         ));
     }